moved tiling button to CanvasSection
