+++ src/components/pages/Generator/CanvasSection/CanvasSection.tsx
@@ -14,6 +14,7 @@ import {drawInvert} from './utils/drawInvert';
 import {saveImage} from './utils/saveImage';
 import {getCtx2dFromRef} from './utils/getCtx2dFromRef';
 import {getCanvasDimensions} from './utils/getCanvasDimensions';
+import { Checkbox } from '@/components/ui/Checkbox/Checkbox';
 
 type Resolution = '1024' | '2048' | '4096' | '8192';
 type PreviewType = 'original' | 'normal' | 'color';
@@ -27,6 +28,8 @@ export function CanvasSection() {
   const [isRendering, setIsRendering] = useState<boolean>(false);
   const [previewType, setPreviewType] = useState<PreviewType>('original');
   const [renderTimeMs, setRenderTimeMs] = useState<number | undefined>();
+  const tilingEnabled = useStore((state) => state.tilingEnabled);
+  const setTilingEnabled = useStore((state) => state.setTilingEnabled);
 
   const canvasRef = useRef<HTMLCanvasElement>(null);
   const canvasOriginalPreviewDataUrl = useRef<string | undefined>(undefined);
@@ -261,6 +264,13 @@ export function CanvasSection() {
         <span className='text-xs text-pink italic'>
           Please note that changing the resolution resets canvas!
         </span>
+      </SubSection>
+      <SubSection title='Tiling'>
+        <Checkbox
+          label='Tileable'
+          isChecked={tilingEnabled}
+          setIsChecked={setTilingEnabled}
+        />
       </SubSection>
       <SubSection title='Inversion' disabled={invertDisabled}>
         <Button disabled={isRendering || invertDisabled} onClick={invert}>
